Upload works, and is connected to Docker Code
