+++ Spring_Server/gs-uploading-files-complete/Dockerfile
@@ -1,0 +1,5 @@
+FROM openjdk
+WORKDIR ..\..\..\..\..\upload-dir
+ADD RevEngDrunnable.jar RevEngDrunnable.jar
+EXPOSE 8000
+CMD java -jar RevEngDrunnable.jar
