+++ Spring_Server/gs-uploading-files-complete/src/main/java/compile_io/Application.java
@@ -1,0 +1,27 @@
+package compile_io;
+
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+
+import compile_io.storage.StorageProperties;
+import compile_io.storage.StorageService;
+
+@SpringBootApplication
+@EnableConfigurationProperties(StorageProperties.class)
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+    @Bean
+    CommandLineRunner init(StorageService storageService) {
+        return (args) -> {
+            storageService.deleteAll();
+            storageService.init();
+        };
+    }
+}
