+++ Spring_Server/gs-uploading-files-complete/src/main/java/compile_io/Greeting.java
@@ -1,0 +1,20 @@
+package compile_io;
+
+public class Greeting {
+
+    private final long id;
+    private final String content;
+
+    public Greeting(long id, String content) {
+        this.id = id;
+        this.content = content;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
