+++ Spring_Server/gs-uploading-files-complete/src/main/java/compile_io/storage/StorageException.java
@@ -1,0 +1,12 @@
+package compile_io.storage;
+
+public class StorageException extends RuntimeException {
+
+    public StorageException(String message) {
+        super(message);
+    }
+
+    public StorageException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
