+++ Spring_Server/gs-uploading-files-complete/src/main/java/compile_io/storage/StorageFileNotFoundException.java
@@ -1,0 +1,12 @@
+package compile_io.storage;
+
+public class StorageFileNotFoundException extends StorageException {
+
+    public StorageFileNotFoundException(String message) {
+        super(message);
+    }
+
+    public StorageFileNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
