+++ Spring_Server/gs-uploading-files-complete/src/main/java/compile_io/storage/StorageProperties.java
@@ -1,0 +1,21 @@
+package compile_io.storage;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties("storage")
+public class StorageProperties {
+
+    /**
+     * Folder location for storing files
+     */
+    private String location = "upload-dir";
+
+    public String getLocation() {
+        return location;
+    }
+
+    public void setLocation(String location) {
+        this.location = location;
+    }
+
+}
