--- Spring_Server/gs-uploading-files-complete/src/main/java/hello/storage/StorageException.java
@@ -1,12 +1,0 @@
-package hello.storage;
-
-public class StorageException extends RuntimeException {
-
-    public StorageException(String message) {
-        super(message);
-    }
-
-    public StorageException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
