--- Spring_Server/gs-uploading-files-complete/src/main/java/hello/storage/StorageFileNotFoundException.java
@@ -1,12 +1,0 @@
-package hello.storage;
-
-public class StorageFileNotFoundException extends StorageException {
-
-    public StorageFileNotFoundException(String message) {
-        super(message);
-    }
-
-    public StorageFileNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
