--- Spring_Server/gs-uploading-files-complete/src/main/java/hello/storage/StorageProperties.java
@@ -1,21 +1,0 @@
-package hello.storage;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties("storage")
-public class StorageProperties {
-
-    /**
-     * Folder location for storing files
-     */
-    private String location = "upload-dir";
-
-    public String getLocation() {
-        return location;
-    }
-
-    public void setLocation(String location) {
-        this.location = location;
-    }
-
-}
