--- Spring_Server/gs-uploading-files-complete/src/main/java/hello/storage/StorageService.java
@@ -1,23 +1,0 @@
-package hello.storage;
-
-import org.springframework.core.io.Resource;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.nio.file.Path;
-import java.util.stream.Stream;
-
-public interface StorageService {
-
-    void init();
-
-    void store(MultipartFile file);
-
-    Stream<Path> loadAll();
-
-    Path load(String filename);
-
-    Resource loadAsResource(String filename);
-
-    void deleteAll();
-
-}
