+++ Spring_Server/gs-uploading-files-complete/src/main/resources/application.properties
@@ -1,3 +1,3 @@
-spring.servlet.multipart.max-file-size=128KB
-spring.servlet.multipart.max-request-size=128KB
+spring.servlet.multipart.max-file-size=100MB
+spring.servlet.multipart.max-request-size=100MB
 spring.http.multipart.enabled=false
