+++ Spring_Server/gs-uploading-files-complete/src/test/java/hello/FileUploadTests.java
@@ -15,6 +15,9 @@ import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
+import compile_io.storage.StorageFileNotFoundException;
+import compile_io.storage.StorageService;
+
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.fileUpload;
@@ -22,9 +25,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import hello.storage.StorageFileNotFoundException;
-import hello.storage.StorageService;
 
 @RunWith(SpringRunner.class)
 @AutoConfigureMockMvc
